Rename the file from ThemeToggleIcon.tsx to theme-toggle-icon.tsx
