+++ apps/web/components/theme-toggle-icon.tsx
@@ -8,9 +8,9 @@ export const ThemeToggleIcon = () => {
 	return (
 		<span className="view-transition-theme-icon">
 			{theme === "dark" ? (
-				<Moon size={16} className="fill-white stroke-gray-3" />
+				<Moon size={17} className="fill-white stroke-gray-3" />
 			) : (
-				<Sun size={16} className="stroke-gray-12" />
+				<Sun size={17} className="stroke-gray-12" />
 			)}
 		</span>
 	);
